refactor: variable name change
