+++ virtual-keyboard/script.js
@@ -1,7 +1,5 @@
 import Keyboard from './modules/keyboard.js';
-import eventsCase from './modules/eventsCase.js';
-
-let cutOrCopy = '';
+import controller from './modules/controller.js';
 
 const keyboard = new Keyboard();
 if (!localStorage.lang) {
@@ -18,33 +16,6 @@ window.addEventListener('keydown', (event) => {
   if (event.shiftKey && event.repeat) {
     keyboard.audio.pause();
   } else keyboard.audio.play();
-  const caretStart = keyboard.keyboardInput.selectionStart;
-  const caretEnd = keyboard.keyboardInput.selectionEnd;
-  const caretPosition = (start, end) => {
-    keyboard.keyboardInput.setSelectionRange(start, end);
-  };
-
-  if (event.ctrlKey && event.code === 'KeyX') {
-    cutOrCopy = keyboard.keyboardInput.value.slice(caretStart, caretEnd);
-    keyboard.keyboardInput.value = `${keyboard.keyboardInput.value.slice(
-      0,
-      caretStart,
-    )}${keyboard.keyboardInput.value.slice(caretEnd)}`;
-    caretPosition(caretStart, caretStart);
-  }
-
-  if (event.ctrlKey && event.code === 'KeyV') {
-    keyboard.keyboardInput.value = `${keyboard.keyboardInput.value.slice(
-      0,
-      caretStart,
-    )}${cutOrCopy}${keyboard.keyboardInput.value.slice(caretEnd)}`;
-    caretPosition(caretStart + cutOrCopy.length, caretStart + cutOrCopy.length);
-  }
-
-  if (event.ctrlKey && event.code === 'KeyC') {
-    cutOrCopy = keyboard.keyboardInput.value.slice(caretStart, caretEnd);
-    caretPosition(caretStart + cutOrCopy.length, caretStart + cutOrCopy.length);
-  }
 
   keyboard.keys.forEach((key) => {
     const char = key.getAttribute('data-code');
@@ -52,7 +23,7 @@ window.addEventListener('keydown', (event) => {
       key.classList.add('keyboard__key--active');
       keyboard.keyboardInput.focus();
 
-      eventsCase(char, key, keyboard, caretStart, caretEnd);
+      controller(char, key, keyboard);
     }
   });
 });
